requirement.txt
pip==20.2.3
pycryptodome==3.15.0
setuptools==49.2.1
